Title: 최솟값 3개, Time: 232ms, Memory: 1MB, Status: Passed - CodeTree

+++ 240415/최솟값 3개/top-3-smallest-number.cpp
@@ -11,7 +11,7 @@ int main() {
     cin >> n;
     for(int i = 0; i < n; i++){
         int a;
-        int sum = 1;
+        long long sum = 1;
         cin >> a;
         pq.push(-a);
         if(pq.size() < 3){
